Small fixes

+++ server/src/middlewares/upload_file.rs
@@ -96,22 +96,28 @@ async fn upload_file(mut payload: Multipart) -> Result<HttpResponse, actix_web::
     // }
 
     while let Ok(Some(field_result)) = payload.try_next().await {
+        let mut field_is_file = true;
         let mut field = field_result;
         let filename = match field.content_disposition().get_filename() {
             Some(cd) => cd.to_string(),
-            None => "unknown".to_string(),
+            None => {
+                field_is_file = false;
+                "unknown".to_string()
+            }
         };
-        
-        // Define the file path where you want to save the uploaded file
-        let file_path = format!("{}/{}", uploads_folder, filename);
-        let file_path_clone = file_path.clone();
-        // Create a new file and write the field data to it
-        let f = web::block(|| File::create(file_path_clone)).await??;
-        while let Some(chunk) = field.try_next().await? {
-            let mut file_pointer_clone = f.try_clone()?;
-            web::block(move || file_pointer_clone.write_all(&chunk)).await??;
+
+        if (field_is_file) {
+            // Define the file path where you want to save the uploaded file
+            let file_path = format!("{}/{}", uploads_folder, filename);
+            let file_path_clone = file_path.clone();
+            // Create a new file and write the field data to it
+            let f = web::block(|| File::create(file_path_clone)).await??;
+            while let Some(chunk) = field.try_next().await? {
+                let mut file_pointer_clone = f.try_clone()?;
+                web::block(move || file_pointer_clone.write_all(&chunk)).await??;
+            }
+            file_paths.push(file_path);
         }
-        file_paths.push(file_path);
     }
     if file_paths.is_empty() {
         Ok(HttpResponse::Ok().body("No file uploaded"))
